Refactored protomek printing package to match case of other protomek package.
